+++ curve25519-dalek/src/backend/serial/u64/field_lemmas/as_nat_lemmas.rs
@@ -1,88 +1,14 @@
 #![allow(unused)]
-use vstd::arithmetic::div_mod::*;
 use vstd::arithmetic::mul::*;
 use vstd::arithmetic::power2::*;
-use vstd::bits::*;
 use vstd::prelude::*;
 
-use super::common_verus::*;
+use super::super::common_verus::div_mod_lemmas::*;
+use super::super::common_verus::mul_lemmas::*;
+use super::super::common_verus::pow_lemmas::*;
+use super::field_core::*;
 
-// Lemmas and spec functions only used in field_verus.rs
-// A lemma should be in this file instead of `common_verus` if:
-//  - It references some constant prominent in `field_verus` (e.g. 51 for bit operations, 2^255 -19)
-//  - It defines or reasons about a spec function relevant only to `field_verus`
 verus! {
-
-// p = 2^255 - 19
-pub open spec fn p() -> nat {
-    (pow2(255) - 19) as nat
-}
-
-// Proof that 2^255 > 19
-pub proof fn pow255_gt_19()
-    ensures
-        pow2(255) > 19
-{
-    lemma2_to64(); // 2^5
-    lemma_pow2_strictly_increases(5, 255);
-}
-
-pub open spec const mask51: u64 = 2251799813685247u64;
-
-// Specialization for b = 51
-pub proof fn lemma_two_factoring_51(k : nat, ai: u64)
-    ensures
-        pow2(k + 51) * ai == pow2(k) * (pow2(51) * ai)
-{
-    lemma_two_factoring(k, 51, ai);
-}
-
-// Evaluation function, given a field element as limbs, reconstruct the nat value it represents.
-pub open spec fn as_nat(limbs: [u64; 5]) -> nat {
-    (limbs[0] as nat) +
-    pow2(51) * (limbs[1] as nat) +
-    pow2(102) * (limbs[2] as nat) +
-    pow2(153) * (limbs[3] as nat) +
-    pow2(204) * (limbs[4] as nat)
-}
-
-// Evaluation function, given a field element as limbs, reconstruct the nat value it represents.
-pub open spec fn as_nat_32_u8(limbs: [u8; 32]) -> nat {
-    // Verus error: `core::iter::range::impl&%15::fold` is not supported
-    // we write them out manually
-    (limbs[0] as nat) +
-    pow2( 1 * 8) * (limbs[ 1] as nat) +
-    pow2( 2 * 8) * (limbs[ 2] as nat) +
-    pow2( 3 * 8) * (limbs[ 3] as nat) +
-    pow2( 4 * 8) * (limbs[ 4] as nat) +
-    pow2( 5 * 8) * (limbs[ 5] as nat) +
-    pow2( 6 * 8) * (limbs[ 6] as nat) +
-    pow2( 7 * 8) * (limbs[ 7] as nat) +
-    pow2( 8 * 8) * (limbs[ 8] as nat) +
-    pow2( 9 * 8) * (limbs[ 9] as nat) +
-    pow2(10 * 8) * (limbs[10] as nat) +
-    pow2(11 * 8) * (limbs[11] as nat) +
-    pow2(12 * 8) * (limbs[12] as nat) +
-    pow2(13 * 8) * (limbs[13] as nat) +
-    pow2(14 * 8) * (limbs[14] as nat) +
-    pow2(15 * 8) * (limbs[15] as nat) +
-    pow2(16 * 8) * (limbs[16] as nat) +
-    pow2(17 * 8) * (limbs[17] as nat) +
-    pow2(18 * 8) * (limbs[18] as nat) +
-    pow2(19 * 8) * (limbs[19] as nat) +
-    pow2(20 * 8) * (limbs[20] as nat) +
-    pow2(21 * 8) * (limbs[21] as nat) +
-    pow2(22 * 8) * (limbs[22] as nat) +
-    pow2(23 * 8) * (limbs[23] as nat) +
-    pow2(24 * 8) * (limbs[24] as nat) +
-    pow2(25 * 8) * (limbs[25] as nat) +
-    pow2(26 * 8) * (limbs[26] as nat) +
-    pow2(27 * 8) * (limbs[27] as nat) +
-    pow2(28 * 8) * (limbs[28] as nat) +
-    pow2(29 * 8) * (limbs[29] as nat) +
-    pow2(30 * 8) * (limbs[30] as nat) +
-    pow2(31 * 8) * (limbs[31] as nat)
-}
 
 // Lemma: If a > b pointwise, then as_nat(a - b) = as_nat(a) - as_nat(b)
 pub proof fn lemma_as_nat_sub(a: [u64;5], b: [u64;5])
@@ -117,70 +43,6 @@ pub proof fn lemma_as_nat_sub(a: [u64;5], b: [u64;5])
         lemma_mul_is_distributive_sub(pow2(3 * 51) as int, a[3] as int, b[3] as int);
         lemma_mul_is_distributive_sub(pow2(4 * 51) as int, a[4] as int, b[4] as int);
     }
-}
-
-pub proof fn lemma_add_then_shift(a: u64, b: u64)
-    requires
-        a < (1u64 << 52),
-        b < (1u64 << 52)
-    ensures
-        (a + b) < (1u64 << 53),
-        ((a + b) as u64 >> 51) < 4
-{
-    lemma2_to64_rest();
-    assert((a + b) < 1u64 << 53) by {
-        assert((1u64 << 52) + (1u64 << 52) == 1u64 << 53) by (compute);
-    }
-    assert(1u64 << 53 == (1u64 << 51) * 4) by (bit_vector);
-    // 0 < b  /\ a < b * c => a/b < c
-    lemma_multiply_divide_lt((a + b) as int, (1u64 << 51) as int, 4int);
-    shift_is_pow2(51);
-    shift_is_pow2(53);
-    assert((a + b) as u64 >> 51 == (a + b) as u64 / (pow2(51) as u64)) by {
-        lemma_u64_shr_is_div((a + b) as u64, 51);
-    }
-    assert(pow2(53) / pow2(51) == 4) by {
-        lemma_pow2_subtracts(51, 53);
-    }
-}
-
-// >> preserves [<=]. Unfortunately, these operations are u128 and
-// we need lemma_u128_shr_is_div.
-pub proof fn lemma_shr_51_le(a: u128, b: u128)
-    requires
-        a <= b
-    ensures
-        (a >> 51) <= (b >> 51)
-{
-    lemma_pow2_pos(51);
-    lemma2_to64_rest(); // pow2(51) value
-    lemma_u128_shr_is_div(a, 51);
-    lemma_u128_shr_is_div(b, 51);
-    lemma_div_is_ordered(a as int, b as int, 51);
-}
-
-// Corollary of above, using the identity (a << x) >> x == a for u64::MAX
-pub proof fn lemma_shr_51_fits_u64(a: u128)
-    requires
-        a <= (u64::MAX as u128) << 51
-    ensures
-        (a >> 51) <= (u64::MAX as u128)
-
-{
-    assert(((u64::MAX as u128) << 51) >> 51 == (u64::MAX as u128)) by (compute);
-    lemma_shr_51_le(a, (u64::MAX as u128) << 51);
-}
-
-// Auxiliary datatype lemma
-// Should work for any k <= 64, but the proofs are convoluted and we can't use BV
-// (x as u64) = x % 2^64, so x = 2^64 * (x / 2^64) + (x as u64). Thus
-// (x as u64) % 2^k = (x as u64) % 2^k, because 2^k | 2^64 * (...) for k <= 64
-pub proof fn lemma_cast_then_mod_51(x: u128)
-    ensures
-        (x as u64) % (pow2(51) as u64) == x % (pow2(51) as u128)
-{
-    lemma2_to64_rest(); // pow2(51 | 64)
-    assert( (x as u64) % 0x8000000000000 == x % 0x8000000000000) by (bit_vector);
 }
 
 // Explicit and mod-p identities for squaring as_nat conversion
@@ -667,184 +529,6 @@ pub proof fn as_nat_k(a: [u64;5], k: u64)
     }
 }
 
-// Auxiliary lemma for reordering terms in the pow2k proof
-pub proof fn lemma_reorder_mul(a: int, b: int)
-    ensures
-        2 * (a * (19 * b)) == 19 * (2 * (a * b))
-{
-    // 2*( a * (19 * b)) = (2 * a) * (19 * b)
-    lemma_mul_is_associative(2, a, 19 * b);
-    // (2 * a) * (19 * b) = (19 * b) * (2 * a) = 19 * (b * (2 * a))
-    lemma_mul_is_associative(19, b, 2 * a);
-    // (b * (2 * a)) = (b * (a * 2)) = 2 * (a * b)
-    lemma_mul_is_associative(b, a, 2);
-}
-
-pub proof fn lemma_mul_sub(ci: int, cj: int, cj_0: int, k: nat)
-    ensures
-        pow2(k) * (ci - pow2(51) * (cj - cj_0)) == pow2(k) * ci - pow2(k + 51) * cj + pow2(k + 51) * cj_0
-{
-    // 2^k (ci - X) = 2^k ci - 2^k X
-    lemma_mul_is_distributive_sub(pow2(k) as int, ci, pow2(51) * (cj - cj_0));
-    // 2^k (2^51 * Y) = (2^k * 2^51) * Y
-    lemma_mul_is_associative(pow2(k) as int, pow2(51) as int, cj - cj_0);
-    // 2^k * 2^51 = 2^(k + 51)
-    lemma_pow2_adds(k, 51);
-    // 2^(k + 51) * (cj - cj_0) = 2^(k + 51) * cj - 2^(k + 51) * cj_0
-    lemma_mul_is_distributive_sub(pow2(k + 51) as int, cj, cj_0);
-}
-
-pub proof fn term_product_bounds(a: [u64; 5], bound: u64)
-    requires
-        19 * bound <= u64::MAX,
-        forall |i: int| 0 <= i < 5 ==> a[i] < bound
-    ensures
-        // c0
-        (a[0] as u128) * (a[0] as u128) < bound * bound,
-        (a[1] as u128) * ((19 * a[4]) as u128) < 19 * (bound * bound),
-        (a[2] as u128) * ((19 * a[3]) as u128) < 19 * (bound * bound),
-        // c1
-        (a[3] as u128) * ((19 * a[3]) as u128) < 19 * (bound * bound),
-        (a[0] as u128) * (a[1] as u128) < (bound * bound),
-        (a[2] as u128) * ((19 * a[4]) as u128) < 19 * (bound * bound),
-        // c2
-        (a[1] as u128) * (a[1] as u128) < (bound * bound),
-        (a[0] as u128) * (a[2] as u128) < (bound * bound),
-        (a[4] as u128) * ((19 * a[3]) as u128) < 19 * (bound * bound),
-        // c3
-        (a[4] as u128) * ((19 * a[4]) as u128) < 19 * (bound * bound),
-        (a[0] as u128) * (a[3] as u128) < (bound * bound),
-        (a[1] as u128) * (a[2] as u128) < (bound * bound),
-        // c4
-        (a[2] as u128) * (a[2] as u128) < (bound * bound),
-        (a[0] as u128) * (a[4] as u128) < (bound * bound),
-        (a[1] as u128) * (a[3] as u128) < (bound * bound)
-{
-    let bound19 = (19 * bound) as u64;
-
-    let a3_19 = (19 * a[3]) as u64;
-    let a4_19 = (19 * a[4]) as u64;
-
-    assert(bound * (19 * bound) == 19 * (bound * bound)) by {
-        lemma_mul_is_associative(19, bound as int, bound as int);
-    }
-
-    // c0
-    lemma_m(a[0], a[0], bound, bound);
-    lemma_m(a[1], a4_19, bound, bound19);
-    lemma_m(a[2], a3_19, bound, bound19);
-
-    // c1
-    lemma_m(a[3], a3_19, bound, bound19);
-    lemma_m(a[0],  a[1], bound, bound);
-    lemma_m(a[2], a4_19, bound, bound19);
-
-    // c2
-    lemma_m(a[1],  a[1], bound, bound);
-    lemma_m(a[0],  a[2], bound, bound);
-    lemma_m(a[4], a3_19, bound, bound19);
-
-    // c3
-    lemma_m(a[4], a4_19, bound, bound19);
-    lemma_m(a[0],  a[3], bound, bound);
-    lemma_m(a[1],  a[2], bound, bound);
-
-    // c4
-    lemma_m(a[2],  a[2], bound, bound);
-    lemma_m(a[0],  a[4], bound, bound);
-    lemma_m(a[1],  a[3], bound, bound);
-}
-
-pub open spec fn c0_0_val(a: [u64; 5]) -> u128 {
-    (a[0] *  a[0] + 2*( a[1] * (19 * a[4]) + a[2] * (19 * a[3]))) as u128
-}
-pub open spec fn c1_0_val(a: [u64; 5]) -> u128 {
-    (a[3] *  (19 * a[3]) + 2 *( a[0] * a[1] + a[2] * (19 * a[4]))) as u128
-}
-pub open spec fn c2_0_val(a: [u64;5]) -> u128{
-    (a[1] *  a[1] + 2*( a[0] *  a[2] + a[4] * (19 * a[3]))) as u128
-}
-pub open spec fn c3_0_val(a: [u64;5]) -> u128{
-    (a[4] * (19 * a[4]) + 2*( a[0] *  a[3] + a[1] * a[2])) as u128
-}
-pub open spec fn c4_0_val(a: [u64;5]) -> u128{
-    (a[2] *  a[2] + 2*( a[0] *  a[4] + a[1] *  a[3])) as u128
-}
-
-pub proof fn c_i_0_bounded(a: [u64; 5], bound: u64)
-    requires
-        19 * bound <= u64::MAX,
-        forall |i: int| 0 <= i < 5 ==> a[i] < bound
-    ensures
-        c0_0_val(a) < 77 * (bound * bound),
-        c1_0_val(a) < 59 * (bound * bound),
-        c2_0_val(a) < 41 * (bound * bound),
-        c3_0_val(a) < 23 * (bound * bound),
-        c4_0_val(a) <  5 * (bound * bound)
-{
-    term_product_bounds(a, bound);
-}
-
-pub open spec fn c0_val(a: [u64; 5]) -> u128 {
-    c0_0_val(a)
-}
-pub open spec fn c1_val(a: [u64; 5]) -> u128 {
-    (c1_0_val(a) + ((c0_val(a) >> 51) as u64) as u128) as u128
-}
-pub open spec fn c2_val(a: [u64;5]) -> u128{
-    (c2_0_val(a) + ((c1_val(a) >> 51) as u64) as u128) as u128
-}
-pub open spec fn c3_val(a: [u64;5]) -> u128{
-    (c3_0_val(a) + ((c2_val(a) >> 51) as u64) as u128) as u128
-}
-pub open spec fn c4_val(a: [u64;5]) -> u128{
-    (c4_0_val(a) + ((c3_val(a) >> 51) as u64) as u128) as u128
-}
-
-pub proof fn c_i_shift_bounded(a: [u64; 5], bound: u64)
-    requires
-        19 * bound <= u64::MAX,
-        77 * (bound * bound) + u64::MAX <= ((u64::MAX as u128) << 51),
-        forall |i: int| 0 <= i < 5 ==> a[i] < bound
-    ensures
-        (c0_val(a) >> 51) <= (u64::MAX as u128),
-        (c1_val(a) >> 51) <= (u64::MAX as u128),
-        (c2_val(a) >> 51) <= (u64::MAX as u128),
-        (c3_val(a) >> 51) <= (u64::MAX as u128),
-        (c4_val(a) >> 51) <= (u64::MAX as u128)
-
-{
-    c_i_0_bounded(a, bound);
-
-    lemma_shr_51_fits_u64(c0_val(a));
-    lemma_shr_51_fits_u64(c1_val(a));
-    lemma_shr_51_fits_u64(c2_val(a));
-    lemma_shr_51_fits_u64(c3_val(a));
-    lemma_shr_51_fits_u64(c4_val(a));
-}
-
-pub open spec fn carry_val(a: [u64; 5]) -> u64 {
-    (c4_val(a) >> 51) as u64
-}
-
-pub open spec fn a0_0_val(a: [u64; 5]) -> u64 {
-    (c0_val(a) as u64) & mask51
-}
-pub open spec fn a1_0_val(a: [u64; 5]) -> u64 {
-    (c1_val(a) as u64) & mask51
-}
-pub open spec fn a2_0_val(a: [u64; 5]) -> u64 {
-    (c2_val(a) as u64) & mask51
-}
-pub open spec fn a3_0_val(a: [u64; 5]) -> u64 {
-    (c3_val(a) as u64) & mask51
-}
-pub open spec fn a4_0_val(a: [u64; 5]) -> u64 {
-    (c4_val(a) as u64) & mask51
-}
-
-
-// dummy, so we can call `verus`
 fn main() {}
 
 }
