+++ curve25519-dalek/src/backend/serial/u64/scalar_lemmas.rs
@@ -18,6 +18,8 @@ use vstd::bits::*;
 use vstd::calc;
 use vstd::prelude::*;
 
+use super::common_verus::shift_lemmas::*;
+
 verus! {
 
 /// Verification: scalar * scalar.invert() ≡ 1 mod L
